Separating Axes and Clones
This update lays the groundwork for the use of the Separating Axis
Theorem with Planars. This is far from a complete implementation; in
fact, I'm not even sure if it works at the moment. (It's been a while).
What's more, the PowerPoint you downloaded has many speedups that still
have yet to be implemented.

This also incudes code that allows Colliders to make clones of
themselves. The cloned Collider has the same dimensions and attributes
as the original, though now that I think of it, it shares a ColliderData
object with the original. That is problematic, and should likely be
tweaked in a future update.