Separating Edges
I wouldn't necessarily refer to this as a finished implementation of the
Separating Axis Theorem, but this is a good milestone before I split off
into another branch. Specifically, this includes code for each of the
colliders to generate their separating edges, which is necessary for SAT
to test cross products.