Animator Revamp and Physics Engine Improvements
The biggest thing coming with this commit is a revamp to the Animator
code! It was messy, but it's been cleaned up and refactored nicely.

There are also a handful of improvments to the physics engine, including
debugging Landscape collision data, Sphere raycasting, and cleaning up
the interface between BroadPhase and NarrowPhase (R.I.P. PairManager).