Input and Transformations
This push includes code for the input manager for the keyboard; the
mouse has yet to be implemented. This facilitates the use of buffers, as
well as nicely wraps up the events of keys being pressed, held, and
released.

This commit also includes some corrections to Quaternion code, and makes
the createTransformation() routines in MatrixMath consisent with one
another. (Some of them were returning incorrect results.)